Moved commands to their own files
