+++ commands/ping.js
@@ -1,0 +1,6 @@
+module.exports = {
+  name: 'ping',
+  execute(msg, args, client) {
+    msg.channel.send(`Ping!`)
+  }
+}
