+++ package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "dependencies": {
     "axios": "^0.20.0",
+    "config.json": "0.0.4",
     "discord.js": "^12.3.1",
     "express": "^4.17.1"
   },
